requirements.txt
ccxt==4.4.53
aiohttp==3.8.5
pandas==2.0.0
numpy==1.24.0
sqlalchemy==2.0.0
python-dotenv==1.0.0
python-binance==1.0.19 